+++ src/main/java/com/generation153/harmonyfree/core/repository/UserRepository.java
@@ -19,6 +19,6 @@ public interface UserRepository extends JpaRepository<User, Long> {
 	
 	boolean existsByFavoriteTracks_Track_Id(Long trackId);
 
-	boolean existsByAuthUserId(Long authUserId);
+	boolean existsByAuthUserId(Integer authUserId);
 
 }
